Add files via upload
This uses the code that was provided by Andres to create a CSV from the api response. I had to add some lines and omit some lines from the original code to get it to work with my API because the format of my JSON output is different than what the code was originally designed for.